requirements/base.txt
Django==4.2
djangorestframework==3.14.0

python-json-logger==2.0.4 # LOGGING

# Library to get transcript from youtube
youtube-transcript-api==0.5.0
pytube==12.1.3

# transcript summary using nlp
spacy==3.5.1

# summary rendering using OpenAI
openai==0.27.4


# corsheaders
django-cors-headers==3.14.0